Оканчательно завез поддержку регистрации (Но пока она юзлесс так как ничего не дает)

+++ src/main/java/com/project/database_2b2t_org_ru/dao/RoleRepository.java
@@ -3,5 +3,5 @@ package com.project.database_2b2t_org_ru.dao;
 import com.project.database_2b2t_org_ru.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Short> {
 }
